Prevent navigation problem for Emacs 30
`treesit-sexp-type-regexp`, `treesit-sentence-type-regexp` and
`treesit-text-type-regexp` has been marked as obsolete on Emacs 30 and
will be removed soon. The new recommended way is to set
`treesit-thing-settings`.

And I found a better way to judge sexp in dart (copied from
`c-ts-mode`), and it worked well for me, I guess we can use it :)